Added code to download, unzip, project, and convert to geopackage for the WBD layers. I also refactored the MP-used function to make system calls. Next steps are subsetting WBD for CONUS and handling system call errors.

+++ lib/acquire_and_preprocess_inputs.py
@@ -21,17 +21,37 @@ from utils.shared_variables import (NHD_URL_PARENT,
                                     NHD_VECTOR_EXTRACTION_PREFIX,
                                     NHD_VECTOR_EXTRACTION_SUFFIX,
                                     PREP_PROJECTION,
-                                    NWM_HYDROFABRIC_URL,
-                                    HEADWATER_NODES)
+                                    NWM_HYDROFABRIC_URL, 
+                                    WBD_NATIONAL_URL)
 
-from utils.shared_functions import pull_file
+from utils.shared_functions import pull_file, run_system_command
     
 
-def project_and_convert_to_gpkg(args):
-    input_gdb, gdb_layer, output_gpkg, projection = args[0], args[1], args[2], args[3]
+def pull_and_prepare_wbd(path_to_saved_data_parent_dir):
     
-    print("Projecting and converting " + gdb_layer)
-    os.system('ogr2ogr -overwrite -progress -f GPKG -t_srs "{projection}" {output_gpkg} {input_gdb} {gdb_layer}'.format(projection=projection, output_gpkg=output_gpkg, input_gdb=input_gdb, gdb_layer=gdb_layer))
+    # -- Acquire and preprocesses NWM data -- #
+    wbd_directory = os.path.join(path_to_saved_data_parent_dir, 'wbd')
+    if not os.path.exists(wbd_directory):
+        os.mkdir(wbd_directory)
+    
+    wbd_gdb_path = os.path.join(wbd_directory, 'WBD_National_GDB.gdb')
+    
+    # Pull and unzip WBD_National_GDB.zip and project and convert to geopackage if not already done previously.
+    if not os.path.exists(wbd_gdb_path):
+        pulled_wbd_zipped_path = os.path.join(wbd_directory, 'WBD_National_GDB.zip')
+        pull_file(WBD_NATIONAL_URL, pulled_wbd_zipped_path)
+        os.system("7za x {pulled_wbd_zipped_path} -o{wbd_directory}".format(pulled_wbd_zipped_path=pulled_wbd_zipped_path, wbd_directory=wbd_directory))
+        
+    # Project and convert to geopackage.
+    procs_list = []        
+    for wbd_layer in ['WBDHU4', 'WBDHU6', 'WBDHU8']:
+        output_gpkg = os.path.join(wbd_directory, wbd_layer + '.gpkg')
+        procs_list.append(['ogr2ogr -overwrite -progress -f GPKG -t_srs "{projection}" {output_gpkg} {wbd_gdb_path} {wbd_layer}'.format(output_gpkg=output_gpkg, wbd_gdb_path=wbd_gdb_path, wbd_layer=wbd_layer, projection=PREP_PROJECTION)])
+   
+    pool = Pool(3)
+    pool.map(run_system_command, procs_list)
+    
+    # Subset WBD layers to CONUS.
 
 
 def pull_and_prepare_nwm_hydrofabric(path_to_saved_data_parent_dir):
@@ -41,35 +61,29 @@ def pull_and_prepare_nwm_hydrofabric(path_to_saved_data_parent_dir):
     if not os.path.exists(nwm_hydrofabric_directory):
         os.mkdir(nwm_hydrofabric_directory)
     pulled_hydrofabric_tar_gz_path = os.path.join(nwm_hydrofabric_directory, 'NWM_channel_hydrofabric.tar.gz')
-    pull_file(NWM_HYDROFABRIC_URL, pulled_hydrofabric_tar_gz_path)
-    
-    os.system("7za x {pulled_hydrofabric_tar_gz_path} -o{nwm_hydrofabric_directory}".format(pulled_hydrofabric_tar_gz_path=pulled_hydrofabric_tar_gz_path, nwm_hydrofabric_directory=nwm_hydrofabric_directory))
-    
-    pulled_hydrofabric_tar_path = pulled_hydrofabric_tar_gz_path.strip('.gz')
-    os.system("7za x {pulled_hydrofabric_tar_path} -o{nwm_hydrofabric_directory}".format(pulled_hydrofabric_tar_path=pulled_hydrofabric_tar_path, nwm_hydrofabric_directory=nwm_hydrofabric_directory))
-    
-    # Delete temporary zip files.
-    os.remove(pulled_hydrofabric_tar_gz_path)
-    os.remove(pulled_hydrofabric_tar_path)
     
     nwm_hydrofabric_gdb = os.path.join(nwm_hydrofabric_directory, 'NWM_v2.0_channel_hydrofabric', 'nwm_v2_0_hydrofabric.gdb')
 
-    procs_list = []
-    for nwm_layer in ['nwm_reaches_conus', 'nwm_waterbodies_conus', 'nwm_catchments_conus']:
-        output_gpkg = os.path.join(nwm_hydrofabric_directory, nwm_layer + '.gpkg')
-        procs_list.append([nwm_hydrofabric_gdb, nwm_layer, output_gpkg, PREP_PROJECTION])        
+    if not os.path.exists(nwm_hydrofabric_gdb):  # Only pull and unzip if the files don't already exist.
+        pull_file(NWM_HYDROFABRIC_URL, pulled_hydrofabric_tar_gz_path)
+        os.system("7za x {pulled_hydrofabric_tar_gz_path} -o{nwm_hydrofabric_directory}".format(pulled_hydrofabric_tar_gz_path=pulled_hydrofabric_tar_gz_path, nwm_hydrofabric_directory=nwm_hydrofabric_directory))
         
-#    # Define paths to and append variables for any temporarily stored datasets for NWM hydrofabric.
-#    headwater_nodes_gdb = os.path.dirname(HEADWATER_NODES)
-#    print(headwater_nodes_gdb)
-#    headwater_nodes_layer = os.path.split(HEADWATER_NODES)[1]
-#    print(headwater_nodes_layer)
-#    headwater_nodes_output_gpkg = os.path.join(nwm_hydrofabric_directory, 'nwm_headwaters.gpkg')
-#    procs_list.append([headwater_nodes_gdb, headwater_nodes_layer, headwater_nodes_output_gpkg, PREP_PROJECTION])
-#    
-    pool = Pool(4)
-    pool.map(project_and_convert_to_gpkg, procs_list)
+        pulled_hydrofabric_tar_path = pulled_hydrofabric_tar_gz_path.strip('.gz')
+        os.system("7za x {pulled_hydrofabric_tar_path} -o{nwm_hydrofabric_directory}".format(pulled_hydrofabric_tar_path=pulled_hydrofabric_tar_path, nwm_hydrofabric_directory=nwm_hydrofabric_directory))
         
+        # Delete temporary zip files.
+        os.remove(pulled_hydrofabric_tar_gz_path)
+        os.remove(pulled_hydrofabric_tar_path)
+        
+        # Project and convert to geopackage.
+        procs_list = []
+        for nwm_layer in ['nwm_reaches_conus', 'nwm_waterbodies_conus', 'nwm_catchments_conus']:
+            output_gpkg = os.path.join(nwm_hydrofabric_directory, nwm_layer + '.gpkg')
+            procs_list.append(['ogr2ogr -overwrite -progress -f GPKG -t_srs "{projection}" {output_gpkg} {nwm_hydrofabric_gdb} {gdb_layer}'.format(projection=PREP_PROJECTION, output_gpkg=output_gpkg, nwm_hydrofabric_gdb=nwm_hydrofabric_gdb, nwm_layer=nwm_layer)])        
+            
+        pool = Pool(4)
+        pool.map(run_system_command, procs_list)
+            
 
 def pull_and_prepare_nhd_data(procs_list):
     
@@ -79,26 +93,29 @@ def pull_and_prepare_nhd_data(procs_list):
     nhd_vector_download_url = procs_list[2]
     nhd_vector_extraction_path = procs_list[3]
     
+    nhd_gdb = nhd_vector_extraction_path.replace('.zip', '.gdb')  # Update extraction path from .zip to .gdb. 
+
     # Download raster and vector, if not already in user's directory (exist check performed by pull_file()).
     nhd_raster_extraction_parent = os.path.dirname(nhd_raster_extraction_path)
     if not os.path.exists(nhd_raster_extraction_parent):
         os.mkdir(nhd_raster_extraction_parent)
-    pull_file(nhd_raster_download_url, nhd_raster_extraction_path)
-    pull_file(nhd_vector_download_url, nhd_vector_extraction_path)
-    
-    # Unzip downloaded GDB.
-    nhd_vector_extraction_parent = os.path.dirname(nhd_vector_extraction_path)
-    huc = os.path.split(nhd_vector_extraction_parent)[1]  # Parse HUC.
-    
-    # Unzip vector and delete zipped file.
-    os.system("7za x {nhd_vector_extraction_path} -o{nhd_vector_extraction_parent}".format(nhd_vector_extraction_path=nhd_vector_extraction_path, nhd_vector_extraction_parent=nhd_vector_extraction_parent))
-    os.remove(nhd_vector_extraction_path)  # Delete the zipped GDB.
-    
-    # -- Project and convert NHDPlusBurnLineEvent and NHDPlusFlowLineVAA vectors to geopackage -- #
-    nhd_gdb = nhd_vector_extraction_path.replace('.zip', '.gdb')  # Update extraction path from .zip to .gdb. 
-    for nhd_layer in ['NHDPlusBurnLineEvent', 'NHDPlusFlowlineVAA']:
-        output_gpkg = os.path.join(nhd_vector_extraction_parent, nhd_layer + huc + '.gpkg')
-        project_and_convert_to_gpkg([nhd_gdb, nhd_layer, output_gpkg, PREP_PROJECTION])  # Use list because function is configured for multiprocessing.
+        
+    if not os.path.exists(nhd_gdb):  # Only pull if not already pulled and processed.
+        pull_file(nhd_raster_download_url, nhd_raster_extraction_path)
+        pull_file(nhd_vector_download_url, nhd_vector_extraction_path)
+        
+        # Unzip downloaded GDB.
+        nhd_vector_extraction_parent = os.path.dirname(nhd_vector_extraction_path)
+        huc = os.path.split(nhd_vector_extraction_parent)[1]  # Parse HUC.
+        
+        # Unzip vector and delete zipped file.
+        os.system("7za x {nhd_vector_extraction_path} -o{nhd_vector_extraction_parent}".format(nhd_vector_extraction_path=nhd_vector_extraction_path, nhd_vector_extraction_parent=nhd_vector_extraction_parent))
+        os.remove(nhd_vector_extraction_path)  # Delete the zipped GDB.
+        
+        # -- Project and convert NHDPlusBurnLineEvent and NHDPlusFlowLineVAA vectors to geopackage -- #
+        for nhd_layer in ['NHDPlusBurnLineEvent', 'NHDPlusFlowlineVAA']:
+            output_gpkg = os.path.join(nhd_vector_extraction_parent, nhd_layer + huc + '.gpkg')
+            run_system_command(['ogr2ogr -overwrite -progress -f GPKG -t_srs "{projection}" {output_gpkg} {nhd_gdb} {nhd_layer}'.format(projection=PREP_PROJECTION, output_gpkg=output_gpkg, nhd_gdb=nhd_gdb, nhd_layer=nhd_layer)])  # Use list because function is configured for multiprocessing.
     
     
 def manage_preprocessing(hucs_of_interest_file_path, path_to_saved_data_parent_dir):
@@ -149,11 +166,15 @@ def manage_preprocessing(hucs_of_interest_file_path, path_to_saved_data_parent_d
         # Append extraction instructions to nhd_procs_list.
         nhd_procs_list.append([nhd_raster_download_url, nhd_raster_extraction_path, nhd_vector_download_url, nhd_vector_extraction_path])
         
-    # Pull NHD Data.
+    # Pull and prepare NHD data.
     pool = Pool(NUM_OF_WORKERS)
     pool.map(pull_and_prepare_nhd_data, nhd_procs_list)
     
+    # Pull and prepare NWM data.
     pull_and_prepare_nwm_hydrofabric(path_to_saved_data_parent_dir)
+
+    # Pull and prepare WBD data.
+    pull_and_prepare_wbd(path_to_saved_data_parent_dir)
 
 
 if __name__ == '__main__':
@@ -161,11 +182,25 @@ if __name__ == '__main__':
     # Get input arguments from command line.
     hucs_of_interest_file_path = sys.argv[1]
     path_to_saved_data_parent_dir = sys.argv[2]  # The parent directory for all saved data.
-
     
     manage_preprocessing(hucs_of_interest_file_path, path_to_saved_data_parent_dir)
 
     
+    # TODO
+    # Get WBD
+    # Create gpkg of HU4, 6, 8 --> create 1 gpkg with sub-layers using original names ogr2ogr -name flag
     
+"""
+on the config file, you can name what you want as long as that variable reflects it `export input_WBD_gdb=$inputDataDir/WBD_National_GDB.gdb
+`
+`export input_WBD_gdb=$inputDataDir/WBD_National_GDB.gdb
+`
+ahhhh why won't it format
+anyways that's it
+
+line 16 in run_by_unit.sh already designates which layer name to pull as they appear in the original GDB file `input_NHD_WBHD_layer=WBDHU$hucUnitLength`
+so as long as you hard-code that as discussed you should be fine
+    
+"""
